pyproject.toml
[tool.poetry]
name = "cleanrlhf"
version = "0.2.1"
description = ""
authors = ["Costa Huang <costa.huang@outlook.com>"]
readme = "README.md"
packages = [
    { include = "cleanrlhf" },
]

[tool.poetry.dependencies]
python = ">=3.9,<3.11.0"
datasets = "^2.8.0"
numpy = "^1.24.1"
torch = "^1.13.1"
jax = "^0.4.8"
flax = "^0.6.8"
optax = "^0.1.4"
wandb = "^0.13.7"
tensorflow = "^2.11.0"
tiktoken = "^0.3.3"
tyro = "^0.4.2"
orbax-checkpoint = "<0.1.8"
colorama = "^0.4.6"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
